requirements.txt
flask
pdfminer.six
python-docx
pdfplumber
